Move implementation of serializer.Marshal to common v5/payload.
This will allow this code to build on freebsd and other architectures
that are not covered by the build tags in `payload_gohai.go`.
