+++ pkg/metadata/v5/payload_gohai.go
@@ -62,19 +62,6 @@ type Payload struct {
 	// TODO: agent_checks
 }
 
-// MarshalJSON serialization a Payload to JSON
-func (p *Payload) MarshalJSON() ([]byte, error) {
-	// use an alias to avoid infinit recursion while serializing
-	type PayloadAlias Payload
-
-	return json.Marshal((*PayloadAlias)(p))
-}
-
-// Marshal not implemented
-func (p *Payload) Marshal() ([]byte, error) {
-	return nil, fmt.Errorf("V5 Payload serialization is not implemented")
-}
-
 // SplitPayload breaks the payload into times number of pieces
 func (p *Payload) SplitPayload(times int) ([]marshaler.Marshaler, error) {
 	// Metadata payloads are analyzed as a whole, so they cannot be split
